Move type guards to separate file and update imports
